requirements.txt
langchain-openai>=0.0.5
openai>=1.0.0
pillow>=9.5.0
typing>=3.7.4.3
